backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
pydantic==2.10.4
python-multipart==0.0.20
faster-whisper==1.1.0
pdfplumber==0.11.0
python-docx==1.1.0
